% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the shiny app of the package EpiSimR}
\usage{
run_app()
}
\value{
No return value, called for side effects. This function launches a Shiny application.
}
\description{
Launch the shiny app of the package EpiSimR
}
\examples{
if (interactive()) {
  library(EpiSimR)
  run_app()
}

}
